Move classes from files to directories
