+++ app/src/main/java/com/kasianov/sergei/omaloma/core/di/dagger/UseCaseModule.kt
@@ -1,16 +1,16 @@
 package com.kasianov.sergei.omaloma.core.di.dagger
-import com.kasianov.sergei.omaloma.domain.usecases.GetAllStoredPublicHolidaysUseCase
-import com.kasianov.sergei.omaloma.domain.usecases.GetAllStoredPublicHolidaysUseCaseImpl
-import com.kasianov.sergei.omaloma.domain.usecases.LoadPublicHolidaysUseCase
-import com.kasianov.sergei.omaloma.domain.usecases.LoadPublicHolidaysUseCaseImpl
-import com.kasianov.sergei.omaloma.domain.usecases.GetStoredPublicHolidayUseCase
-import com.kasianov.sergei.omaloma.domain.usecases.GetStoredPublicHolidayUseCaseImpl
-import com.kasianov.sergei.omaloma.domain.usecases.PerformWikiSearchUseCaseImpl
-import com.kasianov.sergei.omaloma.domain.usecases.PerformWikiSearchUseCase
-import com.kasianov.sergei.omaloma.domain.usecases.GetWikiPageInfoUseCaseImpl
-import com.kasianov.sergei.omaloma.domain.usecases.GetWikiPageInfoUseCase
-import com.kasianov.sergei.omaloma.domain.usecases.GetWikiUrlsListUseCaseImpl
-import com.kasianov.sergei.omaloma.domain.usecases.GetWikiUrlsListUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.publicholusecases.GetAllStoredPublicHolidaysUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.publicholusecases.GetAllStoredPublicHolidaysUseCaseImpl
+import com.kasianov.sergei.omaloma.domain.usecases.publicholusecases.LoadPublicHolidaysUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.publicholusecases.LoadPublicHolidaysUseCaseImpl
+import com.kasianov.sergei.omaloma.domain.usecases.publicholusecases.GetStoredPublicHolidayUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.publicholusecases.GetStoredPublicHolidayUseCaseImpl
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.PerformWikiSearchUseCaseImpl
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.PerformWikiSearchUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.GetWikiPageInfoUseCaseImpl
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.GetWikiPageInfoUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.GetWikiUrlsListUseCaseImpl
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.GetWikiUrlsListUseCase
 import dagger.Binds
 import dagger.Module
