+++ app/src/main/java/com/kasianov/sergei/omaloma/domain/model/publicholidmappers/MapperDBToPublicHoliday.kt
@@ -1,0 +1,18 @@
+package com.kasianov.sergei.omaloma.domain.model.publicholidmappers
+
+import com.kasianov.sergei.omaloma.core.extentions.Mapper
+import com.kasianov.sergei.omaloma.data.database.dto.DBPublicHoliday
+import com.kasianov.sergei.omaloma.domain.model.PublicHoliday
+import javax.inject.Inject
+
+class MapperDBToPublicHoliday @Inject constructor() :
+    Mapper<DBPublicHoliday, PublicHoliday> {
+    override fun mapDto(input: DBPublicHoliday): PublicHoliday {
+        return PublicHoliday(
+            input.name,
+            input.localName,
+            input.dateFormatted,
+            input.countryCode
+        )
+    }
+}
