+++ app/src/main/java/com/kasianov/sergei/omaloma/domain/usecases/publicholusecases/GetAllStoredPublicHolidaysUseCase.kt
@@ -1,0 +1,9 @@
+package com.kasianov.sergei.omaloma.domain.usecases.publicholusecases
+
+import androidx.lifecycle.LiveData
+import com.kasianov.sergei.omaloma.domain.model.PublicHoliday
+
+// Interfaces
+interface GetAllStoredPublicHolidaysUseCase {
+    operator fun invoke(): LiveData<List<PublicHoliday>>
+}
