+++ app/src/main/java/com/kasianov/sergei/omaloma/domain/usecases/publicholusecases/GetStoredPublicHolidayUseCase.kt
@@ -1,0 +1,7 @@
+package com.kasianov.sergei.omaloma.domain.usecases.publicholusecases
+
+import com.kasianov.sergei.omaloma.domain.model.PublicHoliday
+
+interface GetStoredPublicHolidayUseCase {
+    suspend operator fun invoke(name: String): PublicHoliday
+}
