+++ app/src/main/java/com/kasianov/sergei/omaloma/domain/usecases/publicholusecases/LoadPublicHolidaysUseCaseImpl.kt
@@ -1,0 +1,16 @@
+package com.kasianov.sergei.omaloma.domain.usecases.publicholusecases
+
+import com.kasianov.sergei.omaloma.core.extentions.RequestResult
+import com.kasianov.sergei.omaloma.domain.model.PublicHoliday
+import com.kasianov.sergei.omaloma.domain.repository.PublicHolidaysRepo
+import javax.inject.Inject
+
+class LoadPublicHolidaysUseCaseImpl @Inject constructor(
+    private val publicHolidayRepo: PublicHolidaysRepo
+) : LoadPublicHolidaysUseCase {
+
+    override suspend fun invoke(year: String, country: String): RequestResult<List<PublicHoliday>> {
+        return publicHolidayRepo.fetchRemotePublicHolidays(year, country)
+    }
+
+}
