+++ app/src/main/java/com/kasianov/sergei/omaloma/presentation/publicholidays/PubHolDetailsViewModel.kt
@@ -7,9 +7,9 @@ import com.kasianov.sergei.omaloma.core.BaseViewModel
 import com.kasianov.sergei.omaloma.core.extentions.Event
 import com.kasianov.sergei.omaloma.domain.model.PublicHoliday
 import com.kasianov.sergei.omaloma.domain.model.WikiArticle
-import com.kasianov.sergei.omaloma.domain.usecases.GetStoredPublicHolidayUseCase
-import com.kasianov.sergei.omaloma.domain.usecases.GetWikiUrlsListUseCase
-import com.kasianov.sergei.omaloma.domain.usecases.PerformWikiSearchUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.publicholusecases.GetStoredPublicHolidayUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.GetWikiUrlsListUseCase
+import com.kasianov.sergei.omaloma.domain.usecases.wikiusecases.PerformWikiSearchUseCase
 import javax.inject.Inject
 
 class PubHolDetailsViewModel @Inject constructor(
